+++ os/include/timeros/syscall.h
@@ -2,7 +2,7 @@
 #include <stddef.h>
 
 /* syscall */
-uint64_t __SYSCALL(size_t syscall_id, reg_t arg1, reg_t arg2, reg_t arg3);
+int __SYSCALL(size_t syscall_id, reg_t arg1, reg_t arg2, reg_t arg3);
 
 #define __NR_read 63
 #define __NR_write 64
