+++ os/lib/app.c
@@ -1,5 +1,5 @@
 #include <timeros/os.h>
-uint64_t syscall(size_t id, reg_t arg1, reg_t arg2, reg_t arg3) {
+int syscall(size_t id, reg_t arg1, reg_t arg2, reg_t arg3) {
 
     register uintptr_t a0 asm ("a0") = (uintptr_t)(arg1);
     register uintptr_t a1 asm ("a1") = (uintptr_t)(arg2);
